Updated openstack/openstack Project: openstack/neutron  824b4109a2bf4eb26bb4ce44998e60e50379da4d
Check for StaleData errors in retry decorator

If an object is updated in a compare and swap fashion, it
can fail when the change is flushed and raise a
StaleDataError if another process has updated the SQL record
at the same time. We need to catch these with the retry
decorator to restart the update process in a new transaction
with a fresh read of the latest state.

Partially-Implements: bp/push-notifications
Change-Id: I151ffcf47926f5ac66e452974f87e8bc2a906151