Updated openstack/openstack Project: openstack/nova  450c5d33ec9314998534d2b85dc6cdd754b014d2
live-migration ceph: fix typo in ruleset parsing

The rule_id line is formatted as:

"rule_id": 0,

Which means the id integer is the second column of the parsed output,
not the third.

Since a third column doesn't exist, the call to cut returns nothing,
leaving $RULE_ID invalid, which then causes the subsequent ruleset
commands to fail.

We don't see this normally because $CEPH_REPLICAS is set to 1 by
default.

Change-Id: I3255f9421119aafb7407ba7443400bd3ace0d827
Closes-Bug: #1581159