Updated openstack/openstack Project: openstack/nova  e643066fb04e983aa89ee1bc1c11d90db3bbc876
Replace mox with mock for xenapi vm_utils.lookup

The py34 check jobs are failing occasionally when trying to use mox on
the vm_utils.lookup method, seemingly because it is not restored
properly at some point. Rather than chase that down this patch replaces
the usage of mox with mock for patching this method.

Change-Id: I3f36b72ab435dea224eb61d449e13ae80d9b61a8
Closes-Bug: 1590896