Updated openstack/openstack Project: openstack-infra/project-config  a303b62265b4ee865fece5d95c378f10037df25f
Switch the voting for zaqar tempest test

Zaqar's tempest test was introduced in the newton cycle and have
worked fine for a period of time. And the zaqar functional test
may cause the interdependent issue. It's time to switch the
voting from zaqar functional test to tempest for the master branch.

Change-Id: Ica1ff5ac5ceaf675993122243830a5cae0cd8d09