Updated openstack/openstack Project: openstack/neutron  4821196f94d333cb4c310601776f9b2319a6ddf0
Cleanup Liberty to Mitaka upgrade corner cases not necessary anymore

This corner case was handled to consider the version of non reporting
agents which could have the qos extension enabled.

It's not necessary anymore and has been therefore removed.

The devref details have been updated to present tense, as they should be.

Change-Id: Ie40d3f34a46950f5124cf8ef7f486d90513271f4