Updated openstack/openstack Project: openstack-infra/project-config  fb9701a24287ab4ab7842b87e4df576b11dfdec9
Add accessbot for #openstack-fwaas

The Neutron FireWall-as-a-Service team is diverging a bit from the rest
of the Neutron team, and will soon be holding a virtual mid-cycle.  In
order not to clutter the overall #openstack-neutron channel, and to
serve the FWaaS team as it continues to diverge from Neutron, we would
like the #openstack-fwaas channel logged.

Change-Id: I43861e51ae0d2f967121536b9a60d3798ccc1245