Updated openstack/openstack Project: openstack/manila  938ba49e7c48d8e96419657627c8d2d2717a8dd9
Pass context down to ViewBuilder method

The versioned_method decorator helps add/remove fields in
API responses.

The method reads the request object but does not pass down
the context to the called methods.

The context would be helpful to make decisions regarding
whether the fields are specific to an administrator context
or owner/tenant context.

Change-Id: Ic2e2df305362fa3f353fa2dfd75203afc38afb82
Closes-Bug: #1589516