Updated openstack/openstack Project: openstack/zaqar  a2eaac82d4285d3681cf89c9d3ce56e1562ac877
Use debug instead of warning for pipeline implement error

Now pipeline is heavily used in Zaqar for controller, but unfortunately
operators will see a lot of WARNING errors in Zaqar's log because
some of the controllers don't implement one of the methods. That should
be a development details and it's not necessary to expose as WARNING,
otherwise, operators will be scared.

Change-Id: I98c574dc240372627d9e760fc0253c6340a9463f