Updated openstack/openstack Project: openstack-infra/zuul  762f9b688b6edcdb1364fe5a299d490852a6f375
Move a debug logging line by one

If, for some reason, the function name that comes in does not have a :,
which should never happen, we won't see the bogus line in the log
because we split it before we log it.

This seriously doesn't matter - we don't register functions without :'s
in them. But my eyes just can't stop staring at it.

Change-Id: I089ec1969b28367078ce969bfbfc5272f70318e6