Updated openstack/openstack Project: openstack/python-zaqarclient  641e12e202c40aee0e999c965a665fe88968e186
Ensure the float value of api version could work

Now if we pass a float value of api version such as "v1.0" or
"v2.0", client will raise a DriverLoadFailure error. Because
that client can't match a sutiable API entry.

So we should ensure that that some value like "v1.0" could
work as "v1".

Change-Id: I9631a083fd96a535153b6029b95a1d3c8d9966a2
Closes-bug: #1572616