Updated openstack/openstack Project: openstack/tempest  740fe0a8a4a62a93c6406c0ac1b13fc46f6a931b
tempest: use host instead of hypervisor for aggregate test

For the type-1 hypervisors, the service hosts' hostname is different
from the hypervisor's hostname. But a recent commit changed to
use hypervisor's hostname for host to be added to aggregate in the
aggrate tests. This commit is trying to change back to use the
service host but will checking hypervisor's status, so that it can
continue to use service host's hostname but exclude those hosts which
are unavailable.

Change-Id: Ic7b6e99e190bb5f38a3dbaf57a4c690132344c32
Closes-Bug: #1591840