Updated openstack/openstack Project: openstack/oslo.config  b5d65b025c5c9dfca8945a2018c3b2b6e9ccb62e
Required Option

Required options currently are not pushed to the
top of each section, and they are also not marked
as required.  This will mark the option as required
and push it to the top, uncommented.  This will
allow users to know for sure what options are
minimally required to get the service started.

Change-Id: Ic194498ec695d268cce0a26d19f9469ba800412e
Implements: blueprint oslo-generator-required