Updated openstack/openstack Project: openstack/neutron  f98a4fa0a210b711f3b94789b101740ef210d4d3
Check for mysql SAVEPOINT error in retry decorator

Due to lost savepoints in mysql on deadlock errors, we can get
a non-existent savepoint error that is just masking a deadlock
error. This patch adjusts the retry decorator to check for these
savepoint errors as well. See the bug for more details about
the failure.

Closes-Bug: #1590298
Change-Id: I29905817ad7c69986f182ff3f0d58496608cd665