Updated openstack/openstack Project: openstack/ironic  021408417b15c9cb16ee7e11a550d5429e8df7c0
Add cleanwait timeout cleanup process

Previously, if a node in a cleaning state timed out, the timeout
process would not purge certain items from the node's configuration
which resulted in a short circuiting of the logic cleaning being
retried. This was a result of the node clean_step configuration
not being purged upon a timeout occuring.

This change adds a wrapper method around the cleaning failure
error handler to allow the _fail_if_in_state method to call
the error handler, since error handler syntax is not uniform
and the _fail_if_in_state cannot pass arguments.

It also changes the cleaning error handler to permit the error
handler to delete the node clean_step, and cleaning related
driver_internal_info configuration from a node in the event
the node in in CLEANFAIL state.

Change-Id: I9ee5c0b385648c9b7e1d330d5d1af9b2c486a436
Closes-Bug: #1590146