Updated openstack/openstack Project: openstack/python-cinderclient  835747d40fb64bbf78bb331cda0ecf5cdd16de84
Fix argument order for assertEqual to (expected, observed)

assertEqual expects that the arguments provided to it should be
(expected, observed). If a particluar order is kept as a convention,
then it helps to provide a cleaner message to the developer if Unit
Tests fail. The following patch fixes this issue.

TrivialFix

Change-Id: I817a654733e0f3886ca85778a49aa579592f60ff
Closes-Bug: #1259292