Updated openstack/openstack Project: openstack/ironic  336725a37dbb9e9fe1e1c33684143f2f23335a8f
Replace dict.get(key) with dict[key] in unit tests

This replaces dict.get(key) with dict[key] since these are
tests and we should strive to be as clear/exact as we can.
In particular, we shouldn't be doing
    self.assertIsNone(dict.get(key))
since it will pass if 1. the value is None or 2. if the key isn't
in the dict.

Change-Id: I7b399a5f0a53f50a5ebb7febfe4f5c763bed7597
Partial-Bug: #1590808