Updated openstack/openstack Project: openstack/horizon  a5e8a9153bb0f93033e3432cffdc2968f1693553
Correcting detail-row logic to not disrupt styling

The detail-rows on source and flavor transfer tables get styling confused
when items are selected because their detail-row elements remain and thus
confuse the alternate styling.  This patch fixes both of those tables
to exclude the detail-row when the primary row is also excluded.

Test this by (before the patch) selecting, say, the first option from the
available and see how the styling screws up on the remaining ones.  Then
apply the patch and see how everything looks beautiful.

Change-Id: Ia19fd4b47ea08c28dfb95c369b6b3f7271420f04
Closes-Bug: 1582790