Updated openstack/openstack Project: openstack/keystone  ff3b4f6bc7f3f04629a782d6c14c6f4735c12884
Pass a request to controllers instead of a context

Instead of the unformed context dictionary pass a full request object
with access to the context_dict so that existing functions still work.
After this we can replace smaller usages of the context dict with
functions and properties on the request directly.

Change-Id: Ibe822ed7c76a24a7d31d98ce62f873a01e5fb213