Updated openstack/openstack Project: openstack-infra/zuul  89c41fd6402580f366410b73b5d9a419845ef934
Ansible launcher: set keep-hierarchy on log copying

We don't set keep-hierarchy in the log publisher because it only
publishes one file.  But we're tacking on the full ansible
directory to that publisher.  We want keep-hierarchy set so those
files don't just end up in the root.  Modify that in our job
publisher transform.

Change-Id: I188493168ff62f3711fe59fd93e8bd617864e391