Updated openstack/openstack Project: openstack/python-openstackclient  968d8bb662465c3ae5b09d3dd381f89d6541b6bf
Standardize logger usage in volume

self.app.log is the logger in class OpenStackShell,
which should be used to record logs that have nothing
to do with any specific command.

So, use the file logger instead.

This patch also fixes some usage that doesn't follow
rules in:
http://docs.openstack.org/developer/oslo.i18n/guidelines.html
1. add variables to logger as an argument
2. do not wrap variables with str()

Change-Id: I248861a38a4de0412a080046aa7a6f6473c3e082
Implements: blueprint log-usage