Updated openstack/openstack Project: openstack/requirements  a8a1ea968b07f3c5e31b4bad7b305425377ee257
Remove Rally from project.txt

Rally team is willing to maintain requirments.
There are 2 major reasons:

- Rally team would like to use upper constraints for all releases.
  Including the fact that Rally has independent releases
  (~every 3 weeks) it is going to be very hard to remove and readd
  each time requriements job
- Rally team would lkike to use upper contraints of all libs in gates
  to avoid blocking our gates in case of new releases of libs.

Rally a a requriements is not used in any other project (we are on top
of OpenStack) so this change won't affected other projects. 

P.S. We are going to keep our requirements compatible with OpenStack
libs and develop tooling/jobs for that.

Change-Id: I5ee132b450c3112e6d8d8a59b59df2a9716594dc