Updated openstack/openstack Project: openstack/tripleo-heat-templates  4bd42ea849e6d50d16610377b5e89504f5a4f412
Drop extraconfig for neutron-plumgrid.yaml

This patch drops the extraconfig interface in favor
of using the composable services nested stack instead.

The benefit is that it is easier to enable multiple services
(like network and storage backends at the same time) and all
of the plumgrid settings get to live in the same file.

Partially-implements: blueprint composable-services-within-roles

Change-Id: I1c5827e3650a29f7a0258531f84ae0f50f22343d