Updated openstack/openstack Project: openstack/python-openstackclient  ca5e8e6c8540e457a620cc90d321a08e7417de32
Add default IP version and fix help messages for "ip availability list"

There was not a default IP version in "ip availability list"
command, if we used this command without "--ip-version" option,
the output was different from the outputs with the option
"--ip-version 4" and "--ip-version 6" and it is not right.
This patch add default IP version (default is 4) in ``ip
availability list`` command and make this command work properly
without ``--ip-version`` option. And also fix the help message.

Change-Id: Idc08ab6eaf05946eb2ab59bfb3d4497a383d987d
Closes-Bug: #1592761