Updated openstack/openstack Project: openstack/glance  5f48788b001431c3961dd3e1ebf3a1fb2e7ec32b
Add a soft delete functionality for tasks.

Currently there is no mechanism for deleting tasks on regular basis.
This patch adds a new function that is called in the tasks_get_all
function, so that everytime tasks lists are called, the function
checks if tasks in the database have surpassed the expired_at value.
If that is the case, then it sets the deleted value as 1 for all the
expired tasks.

Co-Authored-By: Mike Fedosin <mfedosin@mirantis.com>

Implements https://review.openstack.org/#/c/324648/
Change-Id: I0bde982de948901f6bfbfab9e57cf84891c22052