Updated openstack/openstack Project: openstack/oslo.log  54fbbffeb7e4f2a4e3e7b9bc9130207222dfc65d
Provide a normal method for deprecation warnings.

The deprecated decorator is very useful in situations where you have a
complete function that you want to deprecate. However it doesn't let me
simply trigger a deprecation warning in a code path that i know is
deprecated but doesn't exist in a function.

Create a standard deprecated warning function that has almost the same
signature (what becomes required) and produces the same warning and
fatal_deprecations behaviour for situations where a deprecator is not
useful.

For example:

    if CONF.deprecated_value:
        versionutils.deprecated_warning('deprecated_value',
                                        as_of=versionutils.OCATA)
        # do stuff

Change-Id: Ifd83cd74ac397e7bc6def11a38b46de5dff2acfa
Closes-Bug: #1590223