Updated openstack/openstack Project: openstack/keystone  5ef5f35e5fa8bfb1143e813fc9389a00b2e2db9a
Add mapping rules integration tests

This patch adds the tests for the Mappings API (part of the
Federated Identity API).

The tests added here are not intended to cover all negative and
corner cases, they are rather testing the API in a higher level
and its integration in a working environment.

Change-Id: If245a12a407f960a7ad5f73aa7af717229976ea2