Updated openstack/openstack Project: openstack/cinder  67f3f6d51b8425743c834e07a726af5ee6999fda
Compact Kilo database migrations

This compacts all database migrations up to Kilo into one
initial schema. We are currently creating, updating, dropping,
and modifying the database on installation for every change
since the beginning. This is very inefficient. This will now
limit the number of changes done to the database for new
installations.

Deployments running a pre-Kilo version will need to upgrade to
Kilo before moving to Newton or later. Since Kilo is now EOL,
this shouldn't be an issue.

Change-Id: I39939022de4d2b069f24c11913a5d54b28d7b296