Updated openstack/openstack Project: openstack/cinder  28b575dc2049346ab8558b754d0b1fb4902c4cb8
Clarify glance_api_insecure help text

The help text for the glance_api_insecure config setting
could cause confusion since it refers to using https but
calls it insecure. This clarifies the help text to point
out that this is in regards to SSL cert validation, not
that https is an insecure protocol to use.

Change-Id: Iee81e5392e07a4fb7ad22364e120d954f003f1c1