Updated openstack/openstack Project: openstack/neutron  a2088760cf51af10f0a37262bb7ad36e5b1d0e4a
Allow self-unsubscribing callbacks

This adjusts the notify loop logic to handle the case where
a callback causes a subscription or unsubscription that changes
the subscriber dictionary to change during iteration.

It was just using .items() which solved the problem for py27 but
was not creating an actual copy in py34. This just calls list()
on .items() to make sure we get a list in both cases.

Change-Id: Iee9d675faf30ec714b4f5c77128d8843d545ecfd
Closes-Bug: #1592983