Updated openstack/openstack Project: openstack/oslo.config  91002b6f456561f460562e00e16cd1d46ec2e014
Handle some native python types in config generation

When generating sample config files if the option is not a
oslo_config.Opt with a type_name value we simply put unknown value in
the type help. In particularly the case of keystoneauth those types are
typically standard python types like str.

In cases like this where there is a limited number of well-known types
oslo_config should do it's best to generate a useful type name rather
than have unknowns.

Change-Id: I5c66b50e1f7ab7d61bd818f4e56d376e14789a22