Updated openstack/openstack Project: openstack/oslo.versionedobjects  0bb46f0a2c41c64d1a50debfc78f50627c4c8761
Add objectify decorator for readability

Because it isn't inherently clear what register_if(False) does on top of
a class, a functionally equivalent objectify decorator is added to make
things a little more readable.

The register_if() decorator was also cleaned up a little bit to reuse
the register() decorator. Unit tests were also added for the register()
functions to ensure none of the functionality is changing.

Change-Id: I0bec8e7e587b81dc0484a7b39f0cf57308d3b3d4