Updated openstack/openstack Project: openstack/cinder  1cd90c646a720aa1469ce940c67dd652e7518543
Add test type clarification to devref

We now have unit, functional, and tempest tests in tree. To avoid
confusion, we need to make sure it's clear what each of these test
types are intended to be used for.

This adds a description for each test type and cleans up some of
the documentation around testing in general.

Change-Id: I15bba187889e058adbf03deb4308a41d0a6248e3