Updated openstack/openstack Project: openstack/cinder  0cdb37a6673e5939de230bdda3f4fcf0d27fd398
Remove locks from Pure volume drivers

If we switch to using a retry and are more careful about handling errors
when a host is deleted or volumes are attached we don’t need to use the
lock.

For initialize_connection time when we go to create a host we run the
risk that a terminate_connection running in parallel deleted our host
from underneath us before the volume was attached. To handle that we
just catch the error in that scenario and try to create the host again.

The second issue for initialize_connection is if we try to create a host
and another initialize_connection running in parallel beats us to it. If
the host already exists we can catch the exception for create_host(..)
and retry, which will give us the existing one.

For terminate_connection time we run the risk that between when we
determine a host to be ready for deletion and actually deleting it a
new volume is attached. We can just swallow this error since the Purity
management API will not allow to delete a host with connected volumes.

Change-Id: I8ae75dab279bb983a31595b6af6105a933d5c72c
Closes-Bug: #1588089