Updated openstack/openstack Project: openstack-infra/project-config  c2054f20a63484a66e71168302003ae6139488fe
Add stable branch jobs for tripleo-ci

We don't branch tripleo-ci, but it is used on all of the tripleo
project branches.  We should be running tests against the stable
branches on every proposed change so we don't accidentally break
them.

Only adding the ha and nonha jobs since they test significantly
different features, and 7 jobs is already a lot to get passing
at once.  The upgrades job isn't going to be testing much on ci
changes since there's no upgrade impact to any of those, and we've
still got the master upgrades job to ensure we don't make any
breaking changes to the job as a whole.

Change-Id: I56a84ab20ae8d4e1317497b302348185d7d8d4af