Updated openstack/openstack Project: openstack/neutron  1f12cb4e7db90ff6449915de6df3a52a6272d1c4
Add exponential gap to retry decorator

This sets the retry interval to 100 ms and turns on the
inc_retry_interval flag which doubles the interval on each
failure to our standard retry decorator. This avoids exhausting
all of the retries in a very short window.

With our MAX_RETRIES set to 10, this means it will sleep a cumulative
time of about 8 seconds if all retry attempts fail.

Change-Id: Ife01e8e5444b035cf797e3d575c6799f92a021a5