Updated openstack/openstack Project: openstack/nova  0814461823a9abd739dc036f933671004006c467
Tear down os-disk-config part 1

This removes the portions of the os-disk-config extension which extend
the returned responses by adding attributes to the servers object.

Tests are updated / removed where appropriate.

Follow ups will be done to remove the request handling, as well as the
image response handling.

Part of bp:api-no-more-extensions

Change-Id: I795365ed9d0a786b56fcb18400ceee26ca8f1c96