Updated openstack/openstack Project: openstack/cinder  e72fe062ab868c863c790ed108abe2eadd91dd73
VMware: Deprecate vCenter version 5.1

Log a warning if the vCenter server version is 5.1. The plan
is to enforce minimum version 5.5 (version next to 5.1) in
the Ocata release.

DocImpact
Running Cinder with a VMware vCenter version 5.1 is deprecated.
The minimum required version of vCenter server will be raised
to 5.5 in the 10.0.0 release.

Change-Id: If97412cf58f2ac9ec8c00bf442be9eb770ca5026