Updated openstack/openstack Project: openstack/nova  724bf655704d3ae8995e8a2f81cb5db306a692e1
Fix image meta which is sent to glance v2

This commit fixes recently found issues with converting
images in nova format to glance v2.

1. If property string is unicode we shouldn't convert it
to str.

2. kernel_id and ramdisk_id are custom properties and for
this reason they have to be processed on 'properties' layer
and not with regular attributes.

3. In glance v1 it was possible to set empty string to
kernel_id and ramdisk_id properties. In glance v2 it's
forbidden and that's why we have to convert them to
None values.

Change-Id: I950777c5d0292b4a54d14913611038a18ef41d57
Closes-bug: #1592808