Updated openstack/openstack Project: openstack/glance-specs  6334a9af1f8343b6bd47ee83bc6186320259edc3
Revert "Add distinction layer"

The former commit was merged without PTL approval or an intent
communicated by champion on the proposal. Having such merged proposals
disrupts the direction and roadmap for the project. So, we need to merge
a group of specs that are non-overlapping. We will know what those specs
at the glance virtual midcycle during the week of Jun 13.

This reverts commit af02b822e0c6dd7e05956c3a9fe439e57bbb8498.

Change-Id: I4232dc9a4c0c6681fb693c245aa4d5bc6bb571b3