Updated openstack/openstack Project: openstack-infra/puppet-kerberos  996470c9d14f7f98117cc62961256df7f1caad65
Don't have the server install the client

This prohibits the client being installed some other way.  If we
want to install afs on a kerberos server, it will want to include
the client.  If we don't want to install afs but we do want to
install kerb clients, we will need to add kerberos::client to the
kdc definition.

Change-Id: I6f2a370b521da08184e388ae0f4ceff363af5c0b