Updated openstack/openstack Project: openstack/keystone  355f62bf0cfc54ebe502793d791ddf2c52a4e5e5
Consolidate domain token tests into TokenAPITests

This patch moves a bunch of the domain scoped token tests from TestAuth to
TokenAPITests. It also removes duplicate tests that existed separately in each.
This is an effort to consolidate all the tests from TestAuth into TokenAPITests
so that we have a single class that enforces all of our token behaviors.

Change-Id: If868fb882bca6a4b680c35f76472a18db524ccb5