Updated openstack/openstack Project: openstack/cinder  1816628bb2549703d312456017f2422dea029daf
Nimble driver should return ceiling of volume size

The function manage_existing_get_size returns the floor of volume
size in nimble driver. But in the driver API:"When calculating the
size, round up to the next GB." The nimble driver should be consistent
with the driver API.This patch fixes it by returning the ceiling
of volume size.

Change-Id: If56387fc2391f8d2f0e6b30dc7876e2febb7688d
Closes-Bug: #1586307