Updated openstack/openstack Project: openstack/horizon  021daad8bd21a08fe158fa10f5adc0db4258dd72
Stop memoizing on request arguments in nova api

The request parameter is a django.http.HttpRequest instance object
which contains metadata about the request. This object is a new instance of
django.http.HttpRequest in every view, so memoizing with it memoizes during the
rendering of a single view only, thus preventing multiple api calls only
during the rendering of one view.

Some nova api memoized methods should memoize more persistantly with objects
which do not change after the rendering of each view, like the list of
extensions suported.

However, it is still a good idea to limit memoization of some api calls during
the rendering of a single view, if the query is subject to give different
results (i.e., the flavor_list can change when an administrator adds a flavor
so we need to retreive new data for each view in case a flavor gets added by a
cloud admin).

Implements blueprint: improve-horizon-caching
Co-Authored-By: Daniel Castellanos <luis.daniel.castellanos@intel.com>
Change-Id: I9f87df72cbf149e7b2d1763d7976cb845b1e98f9