Updated openstack/openstack Project: openstack-infra/storyboard-webclient  e63d2c35b9fb0c32e366e4c26af736da9cd799b4
Add a way to make time-moment not use am-time-ago

`am-time-ago` automatically displays a time in the form "3 hours ago".
This behaviour is used by the time-moment directive when displaying
times from within the last day.

This commit adds a way to disable this behaviour and force the actual
time/date to be shown.

Change-Id: Idfc9113cc818e7e95d874513a81278779c2f1e8a