Updated openstack/openstack Project: openstack-infra/storyboard-webclient  7c71fbd303a04ee495ac2af5ca754c53c99221ac
Rework timeline events into a timeline

This patch changes the way timeline events on stories are displayed.

Currently, each event in the timeline is a grey rectangle, leading to
a big wall of grey on stories with a bunch of notifications, which
makes it hard to spot comments easily (and is ugly).

This patch displays the events as an actual "timeline", with icons
representing the events on the timeline and detail provided fairly
minimally. Important events (story creation and comments) stand out
with branding colour, and are easily found at a glance. This also
provides better visual separation of events.

Also, all dates/times in the event timeline are now displayed in the
form "YYYY/MM/DD at HH:mm:ss", rather than "3 hours ago" and similar.

Change-Id: I7d4f9f015919a3eb2ccb0632516b283e6c1f68b0