Updated openstack/openstack Project: openstack/neutron  5d3a6e6aaa577ad002afa7aed260fd7025badd56
Use StringField for tenant_id fields

OVO is going to start enforcing that UUIDFields are UUIDs:
http://docs.openstack.org/developer/oslo.versionedobjects/api/fields.html#oslo_versionedobjects.fields.UUIDField

This is a breaking API change for Neutron since Neutron can be deployed
without Keystone and has historically supported non-UUID tenant IDs. So
we need to use StringField for our tenant_id fields.

The version numbers of the OVO objects are not being bumped because both
UUIDField and StringField are just strings with no validation differences
and they both just serialize to a string. There is no distinguishable
difference between the two until UUIDField starts validating in the future.

Change-Id: I93488701c824a51b5588093b9aba499068ba36e8