Updated openstack/openstack Project: openstack/cinder  a26117f6ae303dac1878b7ad12c93b46201f1c0e
VMware:Persist storage policy ID in vmdk meta-data

The VMDK driver allows the OpenStack admin to specify the storage
tier (vCenter datastore) of a volume (vmdk) using a storage policy
set in the volume type. Currently we use the storage policy only
to determine a datastore where the volume is placed. Some storage
policies, for e.g. those corresponding to vvol and vsan datastores
allow specifying array specific capabilities such as replication.
In order for these capabilities to be applied to the volume, we
should persist the storage policy ID in the vmdk meta-data in
vCenter. This patch fixes this problem by persisting the storage
policy ID in the vmdk meta-data.

Closes-bug: #1593590
Change-Id: I247bf2a072e195c3918f222a9770f8018690db16