Updated openstack/openstack Project: openstack/cinder  e121a876132912d251cd55be9776521683f7c61c
Use cinders test class in cloudbyte unit tests

The cloudbyte unit tests were inheritting directly from
testtols so they weren't getting all the goodness from
the default fixutre setup, most noticeably the logging
setup.  The result was a bunch of noise generated in
the unit test output.

This changes the base class for the cloudbyte unit tests
to pull in cinder.tests instead of testtools directly,
which fixes up the noise.

There are a still a number of places where cloudbyte is
accessing testtools directly, which may or may not be ok.

Change-Id: Ic4cfdc99ba17f0999a382db96020aa583b876020
Closes-Bug: #1593808